+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This build is not intended for submission, but rather for the developers to keep
 - The dice is in preparation.
 - The dice could be rolled (no initiator yet).
 - A button class in in the works.
+- The button class is close to being ready.
 
 # Misonic Project Internal Development Build 2